Version 0.2.3
VimCompile will now try to retrieve the latex master file for multiple file
projects. For this feature to run, vim should be running on the same directory
as the main file.
The main file will be retrieved using *.latexmain if latex-suite is activated,
or %!TEX root=my_main_file.tex in all other case.